refactored out modify and add actions to their own classes to make things modular and also to give some breathing room
