+++ PSA2Tests/PSA2Tests.csproj
@@ -56,6 +56,9 @@
     <None Update="WriteTests\ComparisonData\Actions\Modify\FitMarioModifyCommandWithParametersToCommandWithNoParameters.pac">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <None Update="WriteTests\ComparisonData\Actions\Modify\FitMarioModifyCommandWithPointerParameter.pac">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Update="WriteTests\Data\FitMario.pac">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
